refactor: stat_feature -> char_feature로 이름 변경 (#11)
